server : POC OAI-compat TTS using OuteTTS
