+++ common/arg.cpp
@@ -2216,7 +2216,7 @@ common_params_context common_params_parser_init(common_params & params, llama_ex
             params.vocoder.hf_repo = "ggml-org/WavTokenizer";
             params.vocoder.hf_file = "WavTokenizer-Large-75-F16.gguf";
         }
-    ).set_examples({LLAMA_EXAMPLE_TTS}));
+    ).set_examples({LLAMA_EXAMPLE_TTS, LLAMA_EXAMPLE_SERVER}));
 
     return ctx_arg;
 }
