+++ examples/tts/CMakeLists.txt
@@ -1,5 +1,5 @@
 set(TARGET llama-tts)
-add_executable(${TARGET} tts.cpp)
+add_executable(${TARGET} tts-impl.hpp tts-impl.cpp tts.cpp)
 install(TARGETS ${TARGET} RUNTIME)
 target_link_libraries(${TARGET} PRIVATE llama common ${CMAKE_THREAD_LIBS_INIT})
 target_compile_features(${TARGET} PRIVATE cxx_std_17)
